Add files via upload
Big Basket 🧺

About the dataset 📊
This dataset is basically a big ol' bunch of info about products, all broken down into 10 easy-peasy pieces:

Index: This is just a fancy way of saying it's a unique ID for each item, like a fingerprint in the data world.
Product: The name of the product, just like you'd see it on the website.
Category: The broad group the product falls into, like groceries or home stuff.
Sub-Category: This is like zooming in on the category. So, maybe "groceries" becomes "fruits" or "home stuff" becomes "cleaning supplies."
Brand: Who makes the product? You know, like Nike or that yummy jam brand you love.
Sale Price: How much you gotta pay for it right now.
Market Price: This is kind of like a reference point, showing the usual price for the product.
Type: Another way to classify the product, just for extra organization.
Rating: What other customers think! This is a number showing how much people liked it.
Description: This is where they tell you all the juicy details about the dataset itself, what it includes and how it's put together